+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased (major)
 
+- Run `postcss-cascade-layers` early compared to other PostCSS plugins (breaking)
 - Updated: Support for Node v14+ (major).
 
 ### 1.1.1 (September 17, 2022)